core/db: create client.go, data.go and metadata.go for each database
